backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
pydantic[email]>=2.7.0
email-validator>=2.1.0
supabase>=2.5.0
google-generativeai>=0.7.0
httpx>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
yt-dlp>=2024.4.9
apscheduler>=3.10.4
